chore: configure import
