+++ src/containers/recordForm/medical/DateInput.tsx
@@ -1,9 +1,10 @@
 import "react-datepicker/dist/react-datepicker.css";
 import DatePicker from "react-datepicker";
 import { zhTW } from "date-fns/locale";
+import moment from "moment";
 import { IconCalendarPlus } from "@tabler/icons-react";
 import { forwardRef, useState } from "react";
-import moment from "moment";
+
 import ErrorMessage from "@/components/ErrorMessage";
 
 interface PropsType {
