+++ src/containers/recordForm/medical/ImageInput.tsx
@@ -1,7 +1,8 @@
-import ErrorMessage from "@/components/ErrorMessage";
 import { IconPhoto } from "@tabler/icons-react";
 import Image from "next/image";
 import { forwardRef, useEffect, useState } from "react";
+
+import ErrorMessage from "@/components/ErrorMessage";
 
 interface PropsType {
   message?: string;
